SoftArduino/MinBackend/pyproject.toml
[project]
name = "minbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "pydantic>=2.11.9",
    "pyserial>=3.5",
    "uvicorn>=0.37.0",
]
